- Create repository pattern of risk population
